sandbox: cgroups: move methods to sandbox file
Move sandbox related methods to its own file.
